refactor(vue3.o): vue3.0适配修改
