+++ generator/ts-template-v3/src/services/request.ts
@@ -127,7 +127,6 @@ const axiosResponse = {
       // network状态在app.vue中控制着一个全局的断网提示组件的显示隐藏
       // 关于断网组件中的刷新重新获取数据，会在断网组件中说明
       console.log('断网了~');
-      window.$eventBus.$emit('isBrokenNetwork', true);
     }
   },
 };
