+++ generator/ts-template-v3/types/window.d.ts
@@ -1,0 +1,10 @@
+import type { App } from 'vue';
+
+declare global {
+  declare interface Window {
+    __APP__: App<Element>;
+    LOCAL_CONFIG?: any;
+    LightSDK?: any;
+    $eventBus?: any;
+  }
+}
